Changed nameing convention using .container for smart component
